+++ src/qss.ts
@@ -1,16 +1,17 @@
 export function encode(obj: Record<string, any>, prefix = '') {
-  var k, i, tmp, str='';
+  let k; let i; let tmp; let
+    str = '';
 
   for (k in obj) {
     if ((tmp = obj[k]) !== void 0) {
       if (Array.isArray(tmp)) {
-        for (i=0; i < tmp.length; i++) {
+        for (i = 0; i < tmp.length; i++) {
           str && (str += '&');
-          str += encodeURIComponent(k) + '=' + encodeURIComponent(tmp[i]);
+          str += `${encodeURIComponent(k)}=${encodeURIComponent(tmp[i])}`;
         }
       } else {
         str && (str += '&');
-        str += encodeURIComponent(k) + '=' + encodeURIComponent(tmp);
+        str += `${encodeURIComponent(k)}=${encodeURIComponent(tmp)}`;
       }
     }
   }
@@ -20,21 +21,21 @@ export function encode(obj: Record<string, any>, prefix = '') {
 
 function toValue(mix: string | undefined) {
   if (!mix) return '';
-  var str = decodeURIComponent(mix);
+  const str = decodeURIComponent(mix);
 
-  return str + '';
+  return `${str}`;
 }
 
 export function decode(str: string) {
-  let cursor
-  let k
-  let out: Record<string, any> = {}
-  let arr = str.split('&');
+  let cursor;
+  let k;
+  const out: Record<string, any> = {};
+  const arr = str.split('&');
 
   while (cursor = arr.pop()) {
     cursor = cursor.split('=');
     k = cursor.shift() as string;
-    let outt = out[k]
+    const outt = out[k];
 
     if (outt !== void 0) {
       out[k] = ([] as Array<string>).concat(outt, toValue(cursor.shift()) as string);